feat findTasksByName function

+++ Backend/services/tasksServices.js
@@ -7,6 +7,8 @@ const getAllTasks = async () => {
 };
 
 const createTasks = async (name, quantiy) => {
+  const findTasksName = await models.findTasksByName(name);
+  if (findTasksName.length > 0) throw new Error('Tasks already exists');
   const tasks = await models.createTasks(name, quantiy);
   return tasks;
 };
@@ -17,11 +19,15 @@ const getByIdTasks = async (id) => {
 };
 
 const updateTasks = async (name, quantity, id) => {
+  const tasksId = await models.getByIdTasks(id);
+  if (tasksId.length === 0) throw new Error('Tasks not found');
   const tasks = await models.updateTasks(name, quantity, id);
   return tasks;
 };
 
 const deleteTasks = async (id) => {
+  const tasksId = await models.getByIdTasks(id);
+  if (tasksId.length === 0) throw new Error('Tasks already deleted');
   const tasks = await models.deleteTasks(id);
   return tasks;
 };
